fix: product table
