--- modules/dashboard/src/lib/ui/pages/products/empty/DraftTab.tsx
@@ -1,74 +1,0 @@
-import { Box, useDisclosure } from '@chakra-ui/react';
-import { useSetPaymentPlan } from '@productize-v1.0.0/modules/shared/hooks';
-import { SharedButton, UpgradePlanModal } from '@productize-v1.0.0/modules/shared/ui';
-import { Link } from 'react-router-dom';
-import { ProductCards } from '../../../ProductCards';
-import { EmptyState } from '../../../empty-states/EmptyState';
-
-const DraftTab = () => {
-  const isPremium = useSetPaymentPlan();
-  const { isOpen, onOpen, onClose } = useDisclosure();
-
-  const emptyState = isPremium ? (
-      <>
-          <ProductCards
-              padding={20}
-              bgPos={{ base: `-30%`, md: `initial` }}
-              img={"https://res.cloudinary.com/kingsleysolomon/image/upload/v1699951012/productize/Layer_1_m6pvyg_yz7oz1.png"}
-              bgColor={"purple.100"}
-              bgImg="https://res.cloudinary.com/kingsleysolomon/image/upload/v1699951010/productize/Layer_1_1_uhdwlr_l8njgb.png"
-          />
-          <EmptyState
-              content={{
-                  title: "Oops! Your draft is empty.",
-                  desc: "Lorem ipsum dolor sit amet consectetur. Nec accumsan amet amet velit. Aliquam dictum id pellentesque aenean turpis nisl. Quam etiam.",
-              }}
-              textAlign={{ base: `center` }}
-              showImage={false}
-          >
-              <Link to={`/dashboard/products/new`}>
-                  <SharedButton
-                      text={"Add New Product"}
-                      btnExtras={{
-                          leftIcon: `ei:plus`,
-                      }}
-                      width={"fit-content"}
-                      height={"40px"}
-                      bgColor={"purple.200"}
-                      textColor={"white"}
-                      borderRadius={"4px"}
-                      fontSize={{ base: `sm`, md: `sm` }}
-                  />
-              </Link>
-          </EmptyState>
-      </>
-  ) : (
-      <EmptyState
-          content={{
-              title: "Upgrade your plan to create a plan",
-              desc: "Lorem ipsum dolor sit amet consectetur. Nec accumsan amet amet velit. Aliquam dictum id pellentesque aenean turpis nisl. Quam etiam.",
-              img: `https://res.cloudinary.com/kingsleysolomon/image/upload/v1699951003/productize/Illustration_3_g5iwpj_gurtby.png`,
-          }}
-          textAlign={{ base: `center`, md: `start` }}
-          showImage={true}
-          maxW="100%"
-      >
-          <SharedButton
-              text={"Upgrade Plan"}
-              btnExtras={{
-                  onClick: onOpen,
-              }}
-              width={"fit-content"}
-              height={"48px"}
-              bgColor={"purple.200"}
-              textColor={"white"}
-              borderRadius={"4px"}
-              fontSize={{ base: `sm`, md: `md` }}
-          />
-          <UpgradePlanModal onClose={onClose} isOpen={isOpen} />
-      </EmptyState>
-  );
-  return <Box>{emptyState}</Box>;
-};
-
-export default DraftTab;
